[MOD] Deleting the screen.h and screen.c File
I moved the functions in screen to object and ui, and deleted screen.
The function name change will be done in the 0.3.n version.
