+++ object.c
@@ -1,37 +1,7 @@
 #include <stdio.h>
 #include <windows.h>
-#include "screen.h"
+#include "object.h"
 #include "ui.h"
-
-void cursor(int visible) {
-    HANDLE handle = GetStdHandle(STD_OUTPUT_HANDLE);
-    CONSOLE_CURSOR_INFO cursor = {1, visible};
-    SetConsoleCursorInfo(handle, &cursor);
-}
-
-void textcolor(int x) {
-    HANDLE color;
-    color = GetStdHandle(STD_OUTPUT_HANDLE);
-    SetConsoleTextAttribute(color, x);
-}
-
-int getmaxx()
-{
-	CONSOLE_SCREEN_BUFFER_INFO csbi;
-    GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &csbi);
-    return csbi.srWindow.Right - csbi.srWindow.Left + 1;
-}
-
-int getmaxy()
-{
-	CONSOLE_SCREEN_BUFFER_INFO csbi;
-    GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &csbi);
-	return csbi.srWindow.Bottom - csbi.srWindow.Top + 1;
-}
-
-int getmaxcolor() {
-	return 16;
-}
 
 void gotoxy(int x, int y) {
     HANDLE hConsoleOutput;
@@ -39,17 +9,6 @@ void gotoxy(int x, int y) {
     hConsoleOutput = GetStdHandle(STD_OUTPUT_HANDLE);
     SetConsoleCursorPosition(hConsoleOutput, Cursor_an_Pos);
 }
-
-void clrscr() 
-{
-	system("cls");
-}
-
-void writesat(int x, int y, int c, const char * s) {
-	gotoxy(x, y);
-	textcolor(c);
-	puts(s);
-} 
 
 void writecat(int x, int y, int c, const char ch) {
 	gotoxy(x, y);
